clean up tests into scratch pad
